Define boundaries of proof of concept and minimum viable product
This post defines clear features that can be worked on one at a time.
So, this post is the development roadmap.

Having a development roadmap reduces the number of decisions to make.
(For example, "what should I work on next?" Instead, I *know* what
I'm doing next before I even start a work session).

Watch out: Be flexible with how atomic the roadmap is described.
Try to strike a good balance between building and planning.
A little progress in both is areas consistently will get this *done*.