Update title in front matter to match filename signature
EXPERIMENT
Want the Jekyll Them to automatically parse the changes to titles.
Recognized that the title of each post was the same.But, subtitles
were different. Think name conflict was rendering three posts down
to one.

Expect that updating the title in the front matter of each files
triggers the front matter parser to render the title of the post
in the `Home` view.