feat - Update homepage subtitle
Before, we used "Pause and Ponder." But, this saying is already
attributed to 3Blue1Brown (I knew I heard it sounded familiar).

As far as I can tell, the updating saying is not widely attributed
to anyone or anything.

Prefer the updated subtitle to reduce the chances of conflicting
with other creators.