fix - Correct Now Page front matter to show `Now` in website header
Before, the website header displayed `Foo` even though the `Foo`
page was removed.

The `Now` page included a `date` field in its front matter.
Each of the other pages does not include a `date` field.

So, it's likely that the `Now` page is not being rendered because
the build system expects pages to include only a `layout`, `title`,
and `permalink` field.

Removing the `date` field should fix this bug.